Fixed trying to exhaustively install a missing dependency when
the dependency is locally cloned on the loose Packages folder as
a git repository and the setting "ignore_vcs_packages" is set to
true, and would trigger the error message:

Skipping git package PythonDebugTools since the setting
"ignore_vcs_packages" is set to true

If the file exists on the system, but it is not on the loader? Just
add them. This happens when you develop the dependency on you
computer and the dependency is installed by git.